+++ audiomentations/__init__.py
@@ -1,37 +1,35 @@
-from .augmentations.filters import (
-    PeakingFilter,
-    LowPassFilter,
-    HighPassFilter,
-    LowShelfFilter,
-    HighShelfFilter,
-    FrequencyMask,
-    BandPassFilter,
-    BandStopFilter,
-)
-from .augmentations.spectrogram_transforms import SpecFrequencyMask, SpecChannelShuffle
-from .augmentations.transforms import (
-    AddBackgroundNoise,
-    AddGaussianNoise,
-    AddGaussianSNR,
+from .augmentations.add_background_noise import AddBackgroundNoise
+from .augmentations.add_gaussian_noise import AddGaussianNoise
+from .augmentations.add_gaussian_snr import AddGaussianSNR
+from .augmentations.add_short_noises import AddShortNoises
+from .augmentations.apply_impulse_response import (
     AddImpulseResponse,
     ApplyImpulseResponse,
-    AddShortNoises,
-    Clip,
-    ClippingDistortion,
-    Gain,
-    LoudnessNormalization,
-    Mp3Compression,
-    Normalize,
-    PitchShift,
-    PolarityInversion,
-    Resample,
-    Reverse,
-    Shift,
-    TanhDistortion,
-    TimeMask,
-    TimeStretch,
-    Trim,
 )
+from .augmentations.band_pass_filter import BandPassFilter
+from .augmentations.band_stop_filter import BandStopFilter
+from .augmentations.clip import Clip
+from .augmentations.clipping_distortion import ClippingDistortion
+from .augmentations.frequency_mask import FrequencyMask
+from .augmentations.gain import Gain
+from .augmentations.high_pass_filter import HighPassFilter
+from .augmentations.high_shelf_filter import HighShelfFilter
+from .augmentations.loudness_normalization import LoudnessNormalization
+from .augmentations.low_pass_filter import LowPassFilter
+from .augmentations.low_shelf_filter import LowShelfFilter
+from .augmentations.mp3_compression import Mp3Compression
+from .augmentations.normalize import Normalize
+from .augmentations.peaking_filter import PeakingFilter
+from .augmentations.pitch_shift import PitchShift
+from .augmentations.polarity_inversion import PolarityInversion
+from .augmentations.resample import Resample
+from .augmentations.reverse import Reverse
+from .augmentations.shift import Shift
+from .augmentations.spectrogram_transforms import SpecFrequencyMask, SpecChannelShuffle
+from .augmentations.tanh_distortion import TanhDistortion
+from .augmentations.time_mask import TimeMask
+from .augmentations.time_stretch import TimeStretch
+from .augmentations.trim import Trim
 from .core.composition import Compose, SpecCompose, OneOf, SomeOf
 
 __version__ = "0.21.0"
