+++ audiomentations/augmentations/add_gaussian_noise.py
@@ -1,0 +1,31 @@
+import random
+
+import numpy as np
+
+from audiomentations.core.transforms_interface import BaseWaveformTransform
+
+
+class AddGaussianNoise(BaseWaveformTransform):
+    """Add gaussian noise to the samples"""
+
+    supports_multichannel = True
+
+    def __init__(self, min_amplitude=0.001, max_amplitude=0.015, p=0.5):
+        super().__init__(p)
+        assert min_amplitude > 0.0
+        assert max_amplitude > 0.0
+        assert max_amplitude >= min_amplitude
+        self.min_amplitude = min_amplitude
+        self.max_amplitude = max_amplitude
+
+    def randomize_parameters(self, samples, sample_rate):
+        super().randomize_parameters(samples, sample_rate)
+        if self.parameters["should_apply"]:
+            self.parameters["amplitude"] = random.uniform(
+                self.min_amplitude, self.max_amplitude
+            )
+
+    def apply(self, samples, sample_rate):
+        noise = np.random.randn(*samples.shape).astype(np.float32)
+        samples = samples + self.parameters["amplitude"] * noise
+        return samples
