+++ audiomentations/augmentations/clip.py
@@ -1,0 +1,56 @@
+import functools
+import os
+import random
+import tempfile
+import uuid
+import warnings
+
+import librosa
+import numpy as np
+import sys
+
+from audiomentations.core.audio_loading_utils import load_sound_file
+from audiomentations.core.transforms_interface import BaseWaveformTransform
+from audiomentations.core.utils import (
+    calculate_desired_noise_rms,
+    calculate_rms,
+    calculate_rms_without_silence,
+    convert_decibels_to_amplitude_ratio,
+    convert_float_samples_to_int16,
+    get_file_paths,
+)
+
+
+
+class Clip(BaseWaveformTransform):
+    """
+    Clip audio by specified values. e.g. set a_min=-1.0 and a_max=1.0 to ensure that no
+    samples in the audio exceed that extent. This can be relevant for avoiding integer
+    overflow or underflow (which results in unintended wrap distortion that can sound
+    horrible) when exporting to e.g. 16-bit PCM wav.
+
+    Another way of ensuring that all values stay between -1.0 and 1.0 is to apply
+    PeakNormalization.
+
+    This transform is different from ClippingDistortion in that it takes fixed values
+    for clipping instead of clipping a random percentile of the samples. Arguably, this
+    transform is not very useful for data augmentation. Instead, think of it as a very
+    cheap and harsh limiter (for samples that exceed the allotted extent) that can
+    sometimes be useful at the end of a data augmentation pipeline.
+    """
+
+    supports_multichannel = True
+
+    def __init__(self, a_min=-1.0, a_max=1.0, p=0.5):
+        """
+        :param a_min: float, minimum value for clipping
+        :param a_max: float, maximum value for clipping
+        :param p: The probability of applying this transform
+        """
+        super().__init__(p)
+        assert a_min < a_max
+        self.a_min = a_min
+        self.a_max = a_max
+
+    def apply(self, samples, sample_rate):
+        return np.clip(samples, self.a_min, self.a_max)
