+++ audiomentations/augmentations/frequency_mask.py
@@ -1,0 +1,67 @@
+import random
+import warnings
+
+import numpy as np
+from scipy.signal import butter, sosfilt
+
+from audiomentations.core.transforms_interface import BaseWaveformTransform
+
+
+class FrequencyMask(BaseWaveformTransform):
+    """
+    Mask some frequency band on the spectrogram.
+    Inspired by https://arxiv.org/pdf/1904.08779.pdf
+
+    This transform does basically the same as BandStopFilter
+    """
+
+    supports_multichannel = True
+
+    def __init__(self, min_frequency_band=0.0, max_frequency_band=0.5, p=0.5):
+        """
+        :param min_frequency_band: Minimum bandwidth, float
+        :param max_frequency_band: Maximum bandwidth, float
+        :param p: The probability of applying this transform
+        """
+        super().__init__(p)
+        warnings.warn(
+            "The FrequencyMask class has been deprecated and may be removed in a future"
+            " version of audiomentations. You can use BandStopFilter instead. It has"
+            " different defaults and different parameter randomization that is better"
+            " aligned with human hearing.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        self.min_frequency_band = min_frequency_band
+        self.max_frequency_band = max_frequency_band
+
+    def __butter_bandstop(self, lowcut, highcut, fs, order=5):
+        nyq = 0.5 * fs
+        low = lowcut / nyq
+        high = highcut / nyq
+        sos = butter(order, [low, high], btype="bandstop", output="sos")
+        return sos
+
+    def __butter_bandstop_filter(self, data, lowcut, highcut, fs, order=5):
+        sos = self.__butter_bandstop(lowcut, highcut, fs, order=order)
+        y = sosfilt(sos, data).astype(np.float32)
+        return y
+
+    def randomize_parameters(self, samples, sample_rate):
+        super().randomize_parameters(samples, sample_rate)
+        if self.parameters["should_apply"]:
+            self.parameters["bandwidth"] = random.randint(
+                self.min_frequency_band * sample_rate // 2,
+                self.max_frequency_band * sample_rate // 2,
+            )
+            self.parameters["freq_start"] = random.randint(
+                16, sample_rate // 2 - self.parameters["bandwidth"] - 1
+            )
+
+    def apply(self, samples, sample_rate):
+        bandwidth = self.parameters["bandwidth"]
+        freq_start = self.parameters["freq_start"]
+        samples = self.__butter_bandstop_filter(
+            samples, freq_start, freq_start + bandwidth, sample_rate, order=6
+        )
+        return samples
