+++ audiomentations/augmentations/gain.py
@@ -1,0 +1,39 @@
+import random
+
+from audiomentations.core.transforms_interface import BaseWaveformTransform
+from audiomentations.core.utils import (
+    convert_decibels_to_amplitude_ratio,
+)
+
+
+class Gain(BaseWaveformTransform):
+    """
+    Multiply the audio by a random amplitude factor to reduce or increase the volume. This
+    technique can help a model become somewhat invariant to the overall gain of the input audio.
+
+    Warning: This transform can return samples outside the [-1, 1] range, which may lead to
+    clipping or wrap distortion, depending on what you do with the audio in a later stage.
+    See also https://en.wikipedia.org/wiki/Clipping_(audio)#Digital_clipping
+    """
+
+    supports_multichannel = True
+
+    def __init__(self, min_gain_in_db=-12, max_gain_in_db=12, p=0.5):
+        """
+        :param p: The probability of applying this transform
+        """
+        super().__init__(p)
+        assert min_gain_in_db <= max_gain_in_db
+        self.min_gain_in_db = min_gain_in_db
+        self.max_gain_in_db = max_gain_in_db
+
+    def randomize_parameters(self, samples, sample_rate):
+        super().randomize_parameters(samples, sample_rate)
+        if self.parameters["should_apply"]:
+            self.parameters["amplitude_ratio"] = convert_decibels_to_amplitude_ratio(
+                random.uniform(self.min_gain_in_db, self.max_gain_in_db)
+            )
+
+    def apply(self, samples, sample_rate):
+        return samples * self.parameters["amplitude_ratio"]
+
