+++ audiomentations/augmentations/normalize.py
@@ -1,0 +1,29 @@
+import numpy as np
+
+from audiomentations.core.transforms_interface import BaseWaveformTransform
+
+
+class Normalize(BaseWaveformTransform):
+    """
+    Apply a constant amount of gain, so that highest signal level present in the sound becomes
+    0 dBFS, i.e. the loudest level allowed if all samples must be between -1 and 1. Also known
+    as peak normalization.
+    """
+
+    supports_multichannel = True
+
+    def __init__(self, p=0.5):
+        super().__init__(p)
+
+    def randomize_parameters(self, samples, sample_rate):
+        super().randomize_parameters(samples, sample_rate)
+        if self.parameters["should_apply"]:
+            self.parameters["max_amplitude"] = np.amax(np.abs(samples))
+
+    def apply(self, samples, sample_rate):
+        if self.parameters["max_amplitude"] > 0:
+            normalized_samples = samples / self.parameters["max_amplitude"]
+        else:
+            normalized_samples = samples
+        return normalized_samples
+
