+++ audiomentations/augmentations/trim.py
@@ -1,0 +1,17 @@
+import librosa
+
+from audiomentations.core.transforms_interface import BaseWaveformTransform
+
+
+class Trim(BaseWaveformTransform):
+    """
+    Trim leading and trailing silence from an audio signal using librosa.effects.trim
+    """
+
+    def __init__(self, top_db=20, p=1.0):
+        super().__init__(p)
+        self.top_db = top_db
+
+    def apply(self, samples, sample_rate):
+        samples, lens = librosa.effects.trim(samples, top_db=self.top_db)
+        return samples
