+++ tests/test_polarity_inversion.py
@@ -3,8 +3,7 @@ import unittest
 import numpy as np
 from numpy.testing import assert_almost_equal
 
-from audiomentations.augmentations.transforms import PolarityInversion
-from audiomentations.core.composition import Compose
+from audiomentations import PolarityInversion, Compose
 
 
 class TestPolarityInversion(unittest.TestCase):
